pyproject.toml
[tool.poetry]
name = "XunliBot"
version = "0.1.0"
description = "XunliBot"
authors = []
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
nonebot2 = "2.0.0b5"
websockets = "^10.3"
nonebot-adapter-onebot = "^2.1.3"
nonebot-plugin-manager = "^0.5.9"
nonebot-plugin-apscheduler = "0.1.4"
nonebot-plugin-epicfree = "^0.1.9"
nb-cli = "^0.6.7"
nonebot-plugin-hikarisearch = "^0.1.7"
anti-useragent = "1.0.7"
beautifulsoup4 = "4.11.1"
bs4 = "0.0.1"
lxml = "^4.9.0"
numpy = "^1.23.1"
Pillow = "^9.1.1"
priority = "2.0.0"
pyOpenSSL = "22.0.0"
urllib3 = "1.26.9"
nonebot-plugin-alias = "^0.3.2"
nonebot-plugin-PicMenu = "^0.2"
nonebot-plugin-withdraw = "^0.2.3"
nonebot-plugin-petpet = "^0.3.15"
nonebot-plugin-logo = "^0.2.2"
nonebot-plugin-memes = "^0.3.5"
nonebot-plugin-htmlrender = "^0.2.0.1"
nonebot-plugin-lolmatch = { version = "^0.2.5", python = "<3.11" }
nonebot-plugin-cocdicer = "^0.4.0"

[tool.poetry.dev-dependencies]
nb-cli = "^0.6.0"

[tool.nonebot]
plugins = ["nonebot_plugin_apscheduler",
            "nonebot_plugin_epicfree",
            "nonebot_plugin_manager",
            "nonebot_plugin_hikarisearch",
            "nonebot_plugin_alias",
            "nonebot_plugin_PicMenu",
            "nonebot_plugin_withdraw",
            "nonebot_plugin_petpet",
            "nonebot_plugin_logo",
            "nonebot_plugin_memes",
            "nonebot_plugin_htmlrender",
            "nonebot_plugin_lolmatch",
            "nonebot_plugin_cocdicer"]
plugin_dirs = ["src/plugins"]

[build-system]
requires = ["poetry_core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
